[TM-867] restore view project controller
